[#22] handling exception
