+++ waflog/src/main/kotlin/com/wafflestudio/waflog/domain/user/exception/SeriesNotFoundException.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.waflog.domain.user.exception
+
+import com.wafflestudio.waflog.global.common.exception.DataNotFoundException
+import com.wafflestudio.waflog.global.common.exception.ErrorType
+
+class SeriesNotFoundException(detail: String = "") :
+    DataNotFoundException(ErrorType.DATA_NOT_FOUND, detail)
